factor shake detection logic out into separate class
